refactor(cpu): #43 PR-3 判定クラスタの葉プリミティブを Zig アダプタへ張替
review/VCFパズルが使う判定クラスタ（search/vctHelpers,threatMoves,threatPatterns,vcfCheck, evaluation/winningPatterns,jumpPatterns,threatDetection）の葉プリミティブ呼び出しのみを Zig アダプタへ委譲。高位ロジック（VCT検証/被必勝分類等）は温存。
- checkJumpFour/checkJumpThree/checkStraightFour/点列 → patternsAdapter
- checkForbiddenMove(...).isForbidden → forbiddenAdapter.isForbiddenForBlack（黒専用判定で意味等価、全箇所 color===black ガード内）
- wasm シングルトンを threatAdapter→threatLoader へ移設（preload/get/set は後方互換 re-export）。patternsAdapter が threatLoader を直 import し patternsAdapter→threatAdapter の循環辺を切断（check:circular 緑）。
- reroute で tautology 化（TS側も wasm 経由）し低速化した parity 3本（createsFourThree/findDoubleMiseMoves/vctHelpers）を削除。回帰は patternsParity（raw wasm==TS）+ reviewSnapshot（in-context golden）+ zig native test がカバー。

動作不変: 対局CPUは無干渉（Zig完結・search/未使用）。review判定/採点は Zig searchScore 由来で不変。reviewSnapshot 不変。
